warnings fix

+++ src/gui/desktop/quick_views/photos_model_controller_component.cpp
@@ -155,20 +155,20 @@ int PhotosModelControllerComponent::category() const
         auto it = std::find(cats.cbegin(), cats.cend(), m_categoryFilter);
 
         assert(it != cats.end());
-        return std::distance(cats.cbegin(), it) + 1;
+        return static_cast<int>(std::distance(cats.cbegin(), it)) + 1;
     }
 }
 
 
 float PhotosModelControllerComponent::ratingFrom() const
 {
-    return m_ratingFrom / 2.0;
+    return static_cast<float>(m_ratingFrom) / 2.0f;
 }
 
 
 float PhotosModelControllerComponent::ratingTo() const
 {
-    return m_ratingTo / 2.0;
+    return static_cast<float>(m_ratingTo) / 2.0f;
 }
 
 
@@ -224,7 +224,7 @@ void PhotosModelControllerComponent::setCategory(int category)
 
 void PhotosModelControllerComponent::setRankFrom(float from)
 {
-    m_ratingFrom = from * 2;
+    m_ratingFrom = static_cast<int>(from * 2);
 
     updateModelFilters();
 }
@@ -232,7 +232,7 @@ void PhotosModelControllerComponent::setRankFrom(float from)
 
 void PhotosModelControllerComponent::setRankTo(float to)
 {
-    m_ratingTo = to * 2;
+    m_ratingTo = static_cast<int>(to * 2);
 
     updateModelFilters();
 }
